feat: initial commit with 2022-12 WIL and WIL topic
This is the initial commit for this repo.
Although most of the files are nearly empty,
the journey has begun! Let us keep on the good work.

Introduce repo structure with 'monthly' for monthly WIL and
others for each topic. 'monthly' breaks down into folders
representing each month.

Naming convention for markdown files should be revised.
The prefix 'WIL-' might be unnecessary.